Move Presenter to presenter artifact
